:recycle:  REFACTOR: game.gateway를 module로 분리하고 의존성 관계 수정
