+++ src/game/game.module.ts
@@ -2,7 +2,6 @@ import { Module } from '@nestjs/common';
 import { PreGameModule } from './pre-game/pre-game.module';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { GameHistory } from 'src/common/entities/game-history.entity';
-import { GameGateway } from './game.gateway';
 import { RepositoryModule } from 'src/repository/repository.module';
 import { User } from 'src/common/entities/user.entity';
 import { GameRatingService } from './game-rating/game-rating.service';
@@ -17,6 +16,6 @@ import { GameHistoryModule } from 'src/game-history/game-history.module';
     GameMainModule,
     GameHistoryModule,
   ],
-  providers: [GameGateway, GameRatingService],
+  providers: [GameRatingService],
 })
 export class GameModule {}
