+++ src/game/pre-game/invitation/game-invitation.module.ts
@@ -6,15 +6,15 @@ import { GameInvitationController } from './game-invitation.controller';
 import { GameInvitationGateway } from './game-invitation.gateway';
 import { GameInvitationService } from './game-invitation.service';
 import { GameHistory } from 'src/common/entities/game-history.entity';
-import { GameGateway } from 'src/game/game.gateway';
 import { GameConnectGateway } from '../game-connect.gateway';
 import { GameRatingService } from 'src/game/game-rating/game-rating.service';
 import { GameHistoryModule } from 'src/game-history/game-history.module';
+import { GameMainModule } from 'src/game/game-main/game-main.module';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([User, GameHistory]), RepositoryModule, GameHistoryModule],
+  imports: [TypeOrmModule.forFeature([User, GameHistory]), RepositoryModule, GameMainModule, GameHistoryModule],
   controllers: [GameInvitationController],
-  providers: [GameInvitationService, GameInvitationGateway, GameGateway, GameConnectGateway, GameRatingService],
+  providers: [GameInvitationService, GameInvitationGateway, GameConnectGateway, GameRatingService],
   exports: [GameInvitationGateway],
 })
 export class GameInvitationModule {}
