+++ src/game/pre-game/matching/game-matching.module.ts
@@ -7,20 +7,21 @@ import { RepositoryModule } from 'src/repository/repository.module';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { User } from 'src/common/entities/user.entity';
 import { GameHistory } from 'src/common/entities/game-history.entity';
-import { GameGateway } from 'src/game/game.gateway';
 import { GameConnectGateway } from '../game-connect.gateway';
 import { GameRatingService } from 'src/game/game-rating/game-rating.service';
 import { GameHistoryModule } from 'src/game-history/game-history.module';
+import { GameMainModule } from 'src/game/game-main/game-main.module';
 
 @Module({
   imports: [
     TypeOrmModule.forFeature([User, GameHistory]),
     ScheduleModule.forRoot(),
     RepositoryModule,
+    GameMainModule,
     GameHistoryModule,
   ],
   controllers: [GameMatchingController],
-  providers: [GameMatchingGateway, GameMatchingService, GameGateway, GameConnectGateway, GameRatingService],
+  providers: [GameMatchingGateway, GameMatchingService, GameConnectGateway, GameRatingService],
   exports: [GameMatchingGateway],
 })
 export class GameMatchingModule {}
